requirements.txt
djangorestframework==3.11.0
Django==3.0.7
django-import-export==2.2.0
django-filter==2.3.0
django-cors-headers==3.4.0
psycopg2-binary==2.8.5
Pillow==7.1.2
django-imagekit==4.0.2